requirements.txt
django_extensions
requests
pyserial
#RPi.GPIO
django-cors-headers
paho-mqtt
newsapi-python
